api/requirements.txt
fastapi==0.115.0
psycopg2-binary==2.9.9
uvicorn==0.30.0
python-multipart==0.0.9
sqlalchemy==2.0.35
anthropic==0.39.0
chromadb==0.5.0
pydantic==2.9.0
python-dotenv==1.0.1
httpx==0.27.0
sentence-transformers==3.1.0
reportlab==4.2.0
